Move IFileFingerprint and FileFingerprint
Moved classes to DataAccess/Model directory and namespace.

Updated all package references accordingly.
