+++ FireMothServices/FileScanning/ScanResult.cs
@@ -6,7 +6,7 @@
 namespace RiotClub.FireMoth.Services.FileScanning
 {
     using System.Collections.Generic;
-    using RiotClub.FireMoth.Services.DataAccess;
+    using RiotClub.FireMoth.Services.DataAccess.Model;
 
     /// <summary>
     /// Specifies the result of a file scan operation.
